+++ pyIndego/const.py
@@ -14,20 +14,17 @@ class Methods(Enum):
     HEAD = "HEAD"
 
 
-DEFAULT_URL = "https://api.indego.iot.bosch-si.com/api/v1/"
+DEFAULT_URL = "https://api.indego-cloud.iot.bosch-si.com/api/v1/"
 CONTENT_TYPE_JSON = "application/json"
 CONTENT_TYPE = "Content-Type"
-DEFAULT_BODY = {
-    "accept_tc_id": "202012",
-    "device": "",
-    "os_type": "Android",
-    "os_version": "4.0",
-    "dvc_manuf": "unknown",
-    "dvc_type": "unknown",
-}
 COMMANDS = ("mow", "pause", "returnToDock")
 
-DEFAULT_HEADER = {CONTENT_TYPE: CONTENT_TYPE_JSON}
+DEFAULT_HEADER = {
+    CONTENT_TYPE: CONTENT_TYPE_JSON,
+    # We need to change the user-agent!
+    # The Microsoft Azure proxy seems to block all requests (HTTP 403) for the default 'python-requests' user-agent.
+    "User-Agent": "pyIndego"
+}
 DEFAULT_LOOKUP_VALUE = "Not in database."
 
 DEFAULT_CALENDAR = {